+++ include/eve/module/special/regular/impl/betainc.hpp
@@ -7,22 +7,10 @@
 //==================================================================================================
 #pragma once
 
-#include <eve/concept/value.hpp>
-#include <eve/function/dec.hpp>
-#include <eve/function/exp.hpp>
-#include <eve/function/fma.hpp>
-#include <eve/function/is_ltz.hpp>
-#include <eve/function/inc.hpp>
-#include <eve/function/log.hpp>
-#include <eve/function/log1p.hpp>
-#include <eve/function/lbeta.hpp>
-#include <eve/function/logical_or.hpp>
-#include <eve/function/maxmag.hpp>
-#include <eve/function/oneminus.hpp>
-#include <eve/function/rec.hpp>
-#include <eve/constant/allbits.hpp>
-#include <eve/constant/one.hpp>
-#include <eve/concept/value.hpp>
+#include <eve/module/arithmetic.hpp>
+#include <eve/module/math.hpp>
+#include <eve/module/special/regular/lbeta.hpp>
+#include <eve/module/special/regular/signgam.hpp>
 
 namespace eve::detail
 {
