+++ include/eve/module/special/regular/impl/digamma.hpp
@@ -7,28 +7,10 @@
 //==================================================================================================
 #pragma once
 
+#include <eve/module/arithmetic.hpp>
+#include <eve/module/math.hpp>
+#include <eve/module/special/regular/lgamma.hpp>
 #include <eve/detail/hz_device.hpp>
-#include <eve/concept/value.hpp>
-#include <eve/function/abs.hpp>
-#include <eve/function/add.hpp>
-#include <eve/function/any.hpp>
-#include <eve/function/cotpi.hpp>
-#include <eve/function/dec.hpp>
-#include <eve/function/epsilon.hpp>
-#include <eve/function/exp.hpp>
-#include <eve/function/if_else.hpp>
-#include <eve/function/is_not_nan.hpp>
-#include <eve/function/is_infinite.hpp>
-#include <eve/function/fma.hpp>
-#include <eve/function/fms.hpp>
-#include <eve/function/lgamma.hpp>
-#include <eve/function/log.hpp>
-#include <eve/function/max.hpp>
-#include <eve/function/maximum.hpp>
-#include <eve/function/min.hpp>
-#include <eve/function/rsqrt.hpp>
-#include <eve/constant/one.hpp>
-#include <type_traits>
 
 namespace eve::detail
 {
