+++ include/eve/module/special/regular/impl/tgamma.hpp
@@ -7,30 +7,10 @@
 //==================================================================================================
 #pragma once
 
-#include <eve/concept/value.hpp>
-#include <eve/constant/inf.hpp>
-#include <eve/constant/minf.hpp>
-#include <eve/constant/mone.hpp>
-#include <eve/constant/nan.hpp>
-#include <eve/constant/one.hpp>
-#include <eve/constant/pi.hpp>
-#include <eve/detail/apply_over.hpp>
-#include <eve/detail/implementation.hpp>
-#include <eve/function/abs.hpp>
-#include <eve/function/copysign.hpp>
-#include <eve/function/dec.hpp>
-#include <eve/function/inc.hpp>
-#include <eve/function/is_eqz.hpp>
-#include <eve/function/is_even.hpp>
-#include <eve/function/is_ltz.hpp>
-#include <eve/function/logical_and.hpp>
-#include <eve/function/count_true.hpp>
-#include <eve/function/sinpi.hpp>
-#include <eve/function/stirling.hpp>
-#include <eve/module/real/core/detail/generic/horn.hpp>
-#include <eve/platform.hpp>
-
-#include <type_traits>
+#include <eve/module/arithmetic.hpp>
+#include <eve/module/math.hpp>
+#include <eve/module/special/regular/stirling.hpp>
+#include <eve/detail/hz_device.hpp>
 
 namespace eve::detail
 {
